Update patch set 1
Patch Set 1:

If it is correct, can we at least suppress the log spam? It doesn't seem to be particularly important from what I can tell.

 > Maybe the issue is that this isn't properly handling
 > file_contexts.bin format? It working is still better than it not,
 > but it still doesn't seem to be working properly.

If it is correct, can we at least suppress the log spam? It doesn't seem to be particularly important from what I can tell.

Patch-set: 1